完成药品采购单的相关功能。

+++ MIMS.Entity/Model/PSS_PurchasePlanDetail.cs
@@ -28,7 +28,7 @@ namespace MIMS.Entity.Model
         /// <summary>
         /// PurchaseNum
         /// </summary>		
-        public int PurchaseNum { get; set; }
+        public double PurchaseNum { get; set; }
 
     }
 }
